requirements.txt
python-docx==0.8.11
Pillow==10.0.0 